refactor:회원가입 리펙토링
